// website/WebContent/src/app/generated/api/service-requester/model/offer.ts
/**
 * Service Requester REST Interface
 * Rest interface to communicate with the service requester.
 *
 * OpenAPI spec version: 0.0.2
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CompositionAndOTFProvider } from './compositionAndOTFProvider';
import { PolicyInformation } from './policyInformation';
import { SimplePolicy } from './simplePolicy';


export interface Offer { 
    compositionAndOTFProvider?: CompositionAndOTFProvider;
    compositionID?: string;
    nonFunctionalProperties?: { [key: string]: string; };
    offerScore?: number;
    offerUUID?: string;
    policyInformation?: PolicyInformation;
    simplePolicy?: SimplePolicy;
}